Remove images we use from DockerHub rate limits
DockerHub recently started rate limiting. As a result we are stopping
using DockerHub for high-volume production CI/CD pipelines. This
commit finishes off that work.

This commit creates 3 new Dockerfiles corresponding to the 3 uses of
DockerHub in our pipelines. We want to use GitHub Container Registry,
but these images aren't available there yet.

Alternative is to have something copy these 3 upstream images across
to GitHub Container Registry, but that doesn't fit with the existing
pattern we have in this repo.